Оптимизирует сортировку
- выносит сортировку и список офферов в отдельный модуль
- убирает подписку на стейт сортировки из MainScreen
- исправляет отступ в сортировке
