+++ src/components/sort/sort.jsx
@@ -37,7 +37,7 @@ const Sort = (props) => {
       onFocus={onSortFocus}
       onBlur={onSortBlur}
     >
-      <span className="places__sorting-caption">Sort by</span>
+      <span className="places__sorting-caption">Sort by </span>
       <span className="places__sorting-type" tabIndex="0">
         {SORT_NAMES.find((item) => (item.type === activeSort)).title}
         <svg className="places__sorting-arrow" width="7" height="4">
